dialogue script updates

+++ services/core/api/src/main/kotlin/com/promethist/core/type/Location.kt
@@ -2,9 +2,9 @@ package com.promethist.core.type
 
 import com.fasterxml.jackson.annotation.JsonIgnore
 
-data class Location(val longitude: Float = Float.MIN_VALUE, val latitude: Float = Float.MIN_VALUE) {
+data class Location(val longitude: Double = Double.MIN_VALUE, val latitude: Double = Double.MIN_VALUE) {
     @get:JsonIgnore
-    val isEmpty get() = (longitude == Float.MIN_VALUE && latitude == Float.MIN_VALUE)
+    val isEmpty get() = (longitude == Double.MIN_VALUE && latitude == Double.MIN_VALUE)
     @get:JsonIgnore
     val isNotEmpty get() = !isEmpty
 }
